Sort by document order
I realized that the scope information is off by one in various operations (comma, binary, function call, union). A proper way to compile this that would unlock
a lot more possibilities would be to introduce an IR that is in Administrative Normal Form, but that requires
a lot more work to design this IR. Instead I correct the off-by-one errors by injecting dummy names into
the scope. This work is likely incomplete - step expressions for instance aren't considered yet.

Use the document order information to compile the union properly.