parse function signature, also into ast
The strategy I started Friday was to inspect the arguments of Rust functions and turn them
into a type declaration. This isn't going to work as Rust's introspection isn't powerful enough.
Instead, we will declare the signature of these functions directly with a macro. To this end we had to extend
the facility to parse types and the AST to represent them.