Extract integration tests into a proper integration test.
These were all hiding out in xee_xpath::interpreter::ir_interpret, and it was getting ridiculous.

Instead move them into a separate integration test module. We had to expose various things to the
outside world to make it happen; this API isn't great yet but we'll tackle that another time.