Hook in fancy_regex for fn:matches
Unfortunately while a lot of the tests do work, a lot of them fail. There are a lot of subtle
incompatibilities between the engines. In order to have a compatible engine it looks likely
a new XPath specific regex engine needs to be implemented. This would also help implement xml schema.