Add an explicit SpannedResult
We return this from most public APIs, and modify the ast -> ir and ir -> bytecode
steps to add span information to any errors they generate, so we get better
error messages.

For the query API used by qt3 we've done some hacks for now to work on
the non-spanned level; we need to revise the whole public XPath API anyway
at some point.