Modify how variables are passed.
Previously variables were passed through DynamicContext, but that conflicts with the notion that we don't
want to be recreating a dynamid context each time we do a query, which is especially useful in XSLT.

So we make the variable passing explicit. This has the biggest impact on the test runner.