Just in time document order (#95)
Previously we assigned document order for each node in an annotation hashmap whenever a document got added. This works well for XPath, but XSLT can generate new nodes, and it turns out that XSLT expressions can issue XPath expressions against those in some circumstances, such as when xsl:variable is set to some nodes and later on a select is made against it.

So we've changed the document order annotation system so that annotations are made "just in time"; when requested. There's a caching system and new document order annotations are based on the nearest existing one if possible.

This does introduce a refcell into the system, because we want to treat the cache as immutable.