We introduce a very basic default template rule. (#99)
These are not all the default template rules, even for text only copy, as there are gaps in the pattern implementation.

I've also created the default if `select` is missing in apply-templates.

This makes many hundreds of conformance tests pass.

Note that the conformance test suite still has some flaky tests which sometimes pass and sometimes fail.